Changes to most files to accomodate PHP5.4 depracation of import_request_variables() command. Added transcript file upload function. Added js play button offscreen to player pages for screenreader accessibility.

+++ mediaframe.php
@@ -3,27 +3,25 @@ include "dbconnect.php";
 
 include "playerConfig.php";
 
-//IMPORT VARIABLE assessmentID
-import_request_variables("pg","p_");
+//IMPORT VARIABLE MEDIA ID#
+$mediaID = $_GET['id'];
 
-//passed video id#
-$mediaID = $p_id;
 if(!is_numeric($mediaID)){
 	echo "<div style='width:300px;margin:50 auto;font-size:1.5em;'>Invalid media ID.</div>";
 	exit();
 	}
 
 
-//IS VIDEO IN ALBUM WITH RESTRICTED ACCESS
+//IS VIDEO IN ALBUM WITH RESTRICTED ACCESS?
+$albumPermission = 1;//set flag to public viewing
+//GET ALBUM INFO
 $stmt = $db->prepare("SELECT albummedia.albumID, albums.album, albums.permission FROM albummedia LEFT JOIN albums ON albummedia.albumID = albums.albumID WHERE albummedia.mediaID = :mediaID");
 			$stmt->execute(array(':mediaID'=> $mediaID));
 			$row = $stmt->fetch(PDO::FETCH_ASSOC); 
 			//IF ALBUM PERMISSION RESTRICTED REDIRECT TO HTTPS
 			if(isset($row['permission'])){
 				if($row['permission'] != 'public'){
-					//header("Location: https://" . $server . "/mediasecure.php?id=" . $mediaID);
-					//local install
-					header("Location:mediaframesecure.php?id=" . $mediaID);
+					$albumPermission = 0;//change flag to restricted viewing
 				}
 			}
 
@@ -41,17 +39,19 @@ $size = $row['size'];
 $posterimage = $row['posterimage'];//poster image uploaded or use default?
 $viewcount = $row['viewcount'] + 1;
 
+if($stmt->rowCount() > 0){//video ID exists
 //CHECK VIDEO PERMISSION
-if(($permission == 'public') || ($permission == 'album')){
-	//show video
-}elseif($permission == 'hidden'){
-	echo"<div style='text-align:center;margin-top:150px;'>This video has restricted viewing access.</div>";
-	exit();
-}else{	
+	if($permission == 'public'){
+		//show video
+	}elseif($permission == 'hidden'){
+		echo"<div style='text-align:center;margin-top:150px;'>This video has restricted viewing access.</div>";
+		exit();
+	}elseif(($permission == 'album') && ($albumPermission == 1)){	
+		//show video
+	}else{	
 	//redirect to secure page
-	//header("Location: https://" . $server . "/mediasecure.php?id=" . $mediaID);
-	//local install
 	header("Location:mediaframesecure.php?id=" . $mediaID);
+	}
 }
 	
 //set player size
@@ -70,6 +70,13 @@ $db=null;
 <!-- START OF THE PLAYER EMBEDDING-->
 <script type='text/javascript' src='assets/jwplayer/jwplayer.js'></script>
 <script type="text/javascript">jwplayer.key="<?php echo $playerKey; ?>";</script>
+<style type="text/css">
+	.offscreen {
+		position: absolute;
+		right: 101%;
+		overflow: hidden;
+	}
+	</style>
                                      
 </head>
 
@@ -129,7 +136,11 @@ $db=null;
      		});
     
 			
-	</script>
+</script>
+<!--offscreen play button for screen readers -->
+	<div class="offscreen">
+		<a href="#"  onclick='jwplayer().play()'>Start or Pause Video Playback</a>
+	</div>
 
 <script src="assets/js/jquery.js"></script>
